requirements.txt
gitpython
matplotlib